dbeaver/pro#1905 Moved storing system encoding value to BasApplication class file
