+++ plugins/org.jkiss.dbeaver.ui.app.standalone/src/org/jkiss/dbeaver/ui/app/standalone/DBeaverApplication.java
@@ -271,7 +271,7 @@ public class DBeaverApplication extends DesktopApplicationImpl implements DBPApp
         }
 
         // Encoding initialization
-        GeneralUtils.storeSystemEncoding(System.getProperty(StandardConstants.ENV_FILE_ENCODING));
+        storeSystemEncoding(System.getProperty(StandardConstants.ENV_FILE_ENCODING));
         DBPPreferenceStore store = DBWorkbench.getPlatform().getPreferenceStore();
         if (store != null) {
             GeneralUtils.setFileEncoding(store.getString(DBeaverPreferences.FILE_ENCODING));
